Add bank transaction

+++ backend/bank/src/main/java/com/hasib/bank/dto/UserDto.java
@@ -19,5 +19,5 @@ public class UserDto {
     private long accountNumber;
     private List<Role> roles;
     private boolean isVerified;
-    private long money;
+    private double money;
 }
